Move base protocols to dedicated file, out of Jinja2
